Integration test: wait for Envoy and the upstream server to exit
There appears to be a race condition between the end of one
integration test and the beginning of another.  Envoy can only run one
instance at a time (unless you use the --base-id flag) because it
opens some domain sockets and if a different envoy is running then it
can't create those sockets.

When integration.sh exits (usually because the test program has
finished) it signals Envoy and the upstream http server, which were
started in the background, to exit.  It didn't *wait* until they had
actually exited, though, so if Envoy's shutdown were delayed for some
reason then it could still be running (i.e., still have open domain
sockets) when the next test tried to start Envoy, causing it to fail
with this message:

> [2020-07-29 16:21:17.834][1180][debug][init] [source/common/init/watcher_impl.cc:27] init manager Server destroyed unable to bind domain socket with id=0 (see --base-id option)

https://github.com/envoyproxy/go-control-plane/issues/331